backend/requirements.txt
Django==5.0
djangorestframework==3.16.1
strawberry-graphql==0.287.2
strawberry-graphql-django==0.70.1
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg2-binary==2.9.11
redis==6.4.0
celery[redis]==5.6.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
openai==1.51.0
anthropic==0.34.0
groq==0.9.0
requests==2.32.3
django-filter==24.3
djangorestframework-simplejwt==5.5.1
